The config (/etc) is now stored as a tar archive
Unfortunately due to some changes in the 2.6 kernel,
savedata does not work anymore as expected. When savedata
mounted /dev/ram0 read-only, and dd read the block device
it did not read the actual data. Some cache might not got
invalidated and this caused a problem.

Now cflinux archives the whole /etc to the first partition
of the disk. Before this can be achieved, cflinux must
convert the existing configuration to the new scheme, and
this process may contain bugs, so much testing is needed.